+++ test/index.spec.js
@@ -1,3 +1,5 @@
+import React from 'react';
+import { cleanup, render } from 'react-testing-library';
 import {
   closeActivityForm,
   closeAlertForm,
@@ -63,5 +65,68 @@ describe('Library Index', () => {
 
   it('should expose StoreProvider as a function', () => {
     expect(typeof StoreProvider).toBe('function');
+  });
+});
+
+describe('Component Connect', () => {
+  afterEach(cleanup);
+
+  // eslint-disable-next-line
+  const TestComponent = ({ alerts, total, page }) => (
+    <div>
+      <p data-testid="alerts-count">{typeof alerts}</p>
+      <p data-testid="alerts-total">{total}</p>
+      <p data-testid="alerts-page">{page}</p>
+    </div>
+  );
+
+  it('should render component with states using object accessor', () => {
+    const ConnectedComponent = Connect(TestComponent, {
+      alerts: 'alerts.list',
+      total: 'alerts.total',
+      page: 'alerts.page',
+    });
+
+    const { getByTestId } = render(
+      <StoreProvider>
+        <ConnectedComponent />
+      </StoreProvider>
+    );
+
+    expect(getByTestId('alerts-count').textContent).toBe('object');
+    expect(getByTestId('alerts-total').textContent).toBe('0');
+    expect(getByTestId('alerts-page').textContent).toBe('1');
+  });
+
+  it('should render component with states using functional accessor', () => {
+    const ConnectedComponent = Connect(TestComponent, state => ({
+      alerts: state.alerts.list,
+      total: state.alerts.total,
+      page: state.alerts.page,
+    }));
+
+    const { getByTestId } = render(
+      <StoreProvider>
+        <ConnectedComponent />
+      </StoreProvider>
+    );
+
+    expect(getByTestId('alerts-count').textContent).toBe('object');
+    expect(getByTestId('alerts-total').textContent).toBe('0');
+    expect(getByTestId('alerts-page').textContent).toBe('1');
+  });
+
+  it('should not subscribe to store when accessor is undefined', () => {
+    const ConnectedComponent = Connect(TestComponent);
+
+    const { getByTestId } = render(
+      <StoreProvider>
+        <ConnectedComponent />
+      </StoreProvider>
+    );
+
+    expect(getByTestId('alerts-count').textContent).toBe('undefined');
+    expect(getByTestId('alerts-total').textContent).toBe('');
+    expect(getByTestId('alerts-page').textContent).toBe('');
   });
 });